+++ packages/fiori-docs-embeddings/jest.config.js
@@ -1,0 +1,6 @@
+const config = require('../../jest.base.js');
+
+module.exports = {
+    ...config,
+    modulePathIgnorePatterns: [...config.modulePathIgnorePatterns, '<rootDir>/test/data/']
+};