+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.9.18
+
+### Patch Changes
+
+-   Updated dependencies [9872384]
+    -   @sap-ux/axios-extension@1.22.8
+    -   @sap-ux/btp-utils@1.1.2
+
 ## 0.9.17
 
 ### Patch Changes