+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.4.11
+
+### Patch Changes
+
+-   Updated dependencies [9872384]
+    -   @sap-ux/odata-service-inquirer@2.8.5
+    -   @sap-ux/inquirer-common@0.7.44
+    -   @sap-ux/ui5-config@0.29.5
+    -   @sap-ux/btp-utils@1.1.2
+    -   @sap-ux/abap-deploy-config-sub-generator@0.1.67
+    -   @sap-ux/deploy-config-generator-shared@0.1.36
+    -   @sap-ux/fiori-generator-shared@0.13.17
+    -   @sap-ux/cf-deploy-config-sub-generator@0.2.57
+    -   @sap-ux/project-access@1.32.1
+
+## 0.4.10
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-sub-generator@0.1.66
+
+## 0.4.9
+
+### Patch Changes
+
+-   Updated dependencies [f9b4afe]
+    -   @sap-ux/project-access@1.32.0
+    -   @sap-ux/abap-deploy-config-sub-generator@0.1.65
+    -   @sap-ux/cf-deploy-config-sub-generator@0.2.56
+    -   @sap-ux/fiori-generator-shared@0.13.16
+    -   @sap-ux/odata-service-inquirer@2.8.4
+    -   @sap-ux/deploy-config-generator-shared@0.1.35
+    -   @sap-ux/inquirer-common@0.7.43
+
+## 0.4.8
+
+### Patch Changes
+
+-   Updated dependencies [c385a76]
+    -   @sap-ux/project-access@1.31.0
+    -   @sap-ux/abap-deploy-config-sub-generator@0.1.64
+    -   @sap-ux/cf-deploy-config-sub-generator@0.2.55
+    -   @sap-ux/fiori-generator-shared@0.13.15
+    -   @sap-ux/odata-service-inquirer@2.8.3
+    -   @sap-ux/deploy-config-generator-shared@0.1.34
+    -   @sap-ux/inquirer-common@0.7.42
+
 ## 0.4.7
 
 ### Patch Changes