+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.9.18
+
+### Patch Changes
+
+-   Updated dependencies [9872384]
+    -   @sap-ux/ui5-config@0.29.5
+    -   @sap-ux/project-access@1.32.1
+
+## 0.9.17
+
+### Patch Changes
+
+-   Updated dependencies [f9b4afe]
+    -   @sap-ux/project-access@1.32.0
+
+## 0.9.16
+
+### Patch Changes
+
+-   Updated dependencies [c385a76]
+    -   @sap-ux/project-access@1.31.0
+
 ## 0.9.15
 
 ### Patch Changes