+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/system-access
 
+## 0.6.17
+
+### Patch Changes
+
+-   Updated dependencies [9872384]
+    -   @sap-ux/axios-extension@1.22.8
+    -   @sap-ux/btp-utils@1.1.2
+
 ## 0.6.16
 
 ### Patch Changes