+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.5.12
+
+### Patch Changes
+
+-   Updated dependencies [9872384]
+    -   @sap-ux/ui5-config@0.29.5
+
 ## 1.5.11
 
 ### Patch Changes